Remove inline stylings & unnecessary passing of props to 'CopyClipboardButton'
